+++ recipes/_ldap_config.rb
@@ -7,5 +7,6 @@ template node['grafana']['ini']['auth.ldap']['config_file']['value'] do
   owner 'root'
   group 'root'
   mode '0644'
-  notifies :restart, 'service[grafana-server]', :delayed
+  sensitive true
+  # notifies :restart, 'service[grafana-server]', :delayed
 end