+++ resources/datasource.rb
@@ -1,3 +1,24 @@
+#
+# Cookbook Name:: grafana
+# Resource:: datasource
+#
+# Copyright 2014, Jonathan Tron
+# Copyright 2017, Andrei Skopenko
+# Copyright 2018, Sous Chefs
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
 property :host,           String,   default: 'localhost'
 property :port,           Integer,  default: 3000
 property :admin_user,     String,   default: 'admin'
@@ -107,8 +128,7 @@ action :delete do
 end
 
 def _legacy_http_semantic
-  return false if node['grafana']['version'] == 'latest'
-  Gem::Version.new(node['grafana']['version']) < Gem::Version.new('2.0.3')
+  false
 end
 
 def _check_org!(datasource, orgs)